# api/docs/swagger.yaml
definitions:
...
  models.CreateInvestor:
    properties:
      name:
        type: string
    type: object
  models.CreateOrder:
    properties:
      car_id:
        type: string
      client_id:
        type: string
      day_count:
        type: integer
      give_km:
        type: integer
      paid_price:
        type: number
      recieve_km:
        type: integer
      total_price:
        type: number
    type: object
  models.Empty:
    type: object
...
  models.GetListInvestorResponse:
    properties:
      count:
        type: integer
      investors:
        items:
          $ref: '#/definitions/models.Investor'
        type: array
    type: object
  models.GetListInvestorResponseFoyda:
    properties:
      count:
        type: integer
      investors:
        items:
          $ref: '#/definitions/models.InvestorFoyda'
        type: array
    type: object
  models.GetListOrderResponse:
    properties:
      count:
        type: integer
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
    type: object
  models.Investor:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.InvestorFoyda:
    properties:
      investor_id:
        type: string
      name:
        type: string
      sum:
        type: number
    type: object
  models.Order:
    properties:
      car_id:
        type: string
      client_id:
        type: string
      created_at:
        type: string
      day_count:
        type: integer
      give_km:
        type: integer
      id:
        type: string
      paid_price:
        type: number
      recieve_km:
        type: integer
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
...
  models.UpdateInvestorSwag:
    properties:
      name:
        type: string
    type: object
  models.UpdateOrderSwag:
    properties:
      car_id:
        type: string
      client_id:
        type: string
      day_count:
        type: integer
      give_km:
        type: integer
      paid_price:
        type: number
      recieve_km:
        type: integer
      status:
        type: string
      total_price:
        type: number
    type: object
  models.UpdateStatusSwag:
    properties:
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /investor:
    get:
      consumes:
      - application/json
      description: Get List Investor
      operationId: get_list_Investor
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetInvestorListBody
          schema:
            $ref: '#/definitions/models.GetListInvestorResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Investor
      tags:
      - Investor
    post:
      consumes:
      - application/json
      description: Create Investor
      operationId: create_Investor
      parameters:
      - description: CreateInvestorRequestBody
        in: body
        name: Investor
        required: true
        schema:
          $ref: '#/definitions/models.CreateInvestor'
      produces:
      - application/json
      responses:
        "201":
          description: GetInvestorBody
          schema:
            $ref: '#/definitions/models.Investor'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Investor
      tags:
      - Investor
  /investor-share:
    get:
      consumes:
      - application/json
      description: Get List Investor
      operationId: get_list_Investor
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetInvestorListBody
          schema:
            $ref: '#/definitions/models.GetListInvestorResponseFoyda'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Investor
      tags:
      - Investor Share
  /investor/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Investor
      operationId: delete_Investor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteInvestorBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Investor
      tags:
      - Investor
    get:
      consumes:
      - application/json
      description: Get By ID Investor
      operationId: get_by_id_investor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetInvestorBody
          schema:
            $ref: '#/definitions/models.Investor'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Investor
      tags:
      - Investor
    put:
      consumes:
      - application/json
      description: Update Investor
      operationId: update_Investor
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateInvestorRequestBody
        in: body
        name: Investor
        required: true
        schema:
          $ref: '#/definitions/models.UpdateInvestorSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateInvestorBody
          schema:
            $ref: '#/definitions/models.Investor'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Investor
      tags:
      - Investor
  /order:
    get:
      consumes:
      - application/json
      description: Get List Order
      operationId: get_list_order
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: GetOrderListBody
          schema:
            $ref: '#/definitions/models.GetListOrderResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Order
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Create Order
      operationId: create_order
      parameters:
      - description: CreateOrderRequestBody
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrder'
      produces:
      - application/json
      responses:
        "201":
          description: GetOrderBody
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Order
      tags:
      - Order
  /order/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Order
      operationId: delete_order
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteOrderBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Order
      tags:
      - Order
    get:
      consumes:
      - application/json
      description: Get By ID Order
      operationId: get_by_id_order
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetOrderBody
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Order
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: Update Order
      operationId: update_order
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateOrderRequestBody
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOrderSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateOrderBody
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Order
      tags:
      - Order
  /orderstatus/{id}:
    put:
      consumes:
      - application/json
      description: Update Order Status
      operationId: update_order_status
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateStatusRequestBody
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStatusSwag'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateStatusBody
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Order Status
      tags:
      - Order Status
swagger: "2.0"